refactor: dropdown, sheet, toggle component separation
