+++ my-monorepo/packages/ui/src/components/toggle/ToggleClass.ts
@@ -1,0 +1,15 @@
+export const sizeClass = {
+  xsmall: "px-[7px] py-[1px] text-[12px]",
+  small: "px-[10px] py-[3px] text-[14px]",
+  medium: "px-[12px] py-[5px] text-[16px]",
+  large: "px-[14px] py-[7px] text-[16px]",
+  xlarge: "px-[16px] py-[7px] text-[18px]",
+};
+
+export const radiusClass = {
+  none: "rounded-none",
+  small: "rounded-[5px]",
+  medium: "rounded-[8px]",
+  large: "rounded-[12px]",
+  full: "rounded-full",
+};
